requirements.txt
python-telegram-bot==13.15
googletrans==4.0.0-rc1


pytesseract
Pillow

